requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.7.0
astrapy==1.5.2
beautifulsoup4==4.12.3
cachetools==5.5.0
cassandra-driver==3.29.2
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
fastapi==0.115.6
ffmpeg-python==0.2.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2024.12.0
future==1.0.0
geomet==0.2.1.post1
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
groq==0.15.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.27.1
hyperframe==6.0.1
idna==3.10
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.2.1
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
packaging==24.2
pillow==11.1.0
proto-plus==1.25.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pymongo==4.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
sympy==1.13.1
tabulate==0.9.0
threadpoolctl==3.5.0
tokenizers==0.21.0
toml==0.10.2
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
triton==3.1.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uuid6==2024.7.10
uvicorn==0.34.0
wordcloud==1.9.4
xformers==0.0.29.post1
yt-dlp==2025.1.15
